# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Decline": {
        "additionalProperties": false,
        "properties": {
          "reason": {
            "maxLength": 1000,
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "Decline",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "qarz__interface__payment_notices_api__Accept": {
        "additionalProperties": false,
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          }
        },
        "title": "Accept",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Qarz Daftari",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/me/accounts/{link_id}/payment-notices": {
      "post": {
        "operationId": "me_accounts_payment_notices_send_api_v1_me_accounts__link_id__payment_notices_post",
        "parameters": [
          {
            "in": "path",
            "name": "link_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Link Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Me Accounts Payment Notices Send Api V1 Me Accounts  Link Id  Payment Notices Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Me.Accounts.Payment Notices.Send"
      }
    },
...
    "/api/v1/shops/{shop_id}/payment-notices": {
      "get": {
        "operationId": "payment_notices_list_api_v1_shops__shop_id__payment_notices_get",
        "parameters": [
          {
            "in": "path",
            "name": "shop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shop Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Payment Notices List Api V1 Shops  Shop Id  Payment Notices Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Payment Notices.List"
      }
    },
    "/api/v1/shops/{shop_id}/payment-notices/{notice_id}/accept": {
      "post": {
        "operationId": "payment_notices_accept_api_v1_shops__shop_id__payment_notices__notice_id__accept_post",
        "parameters": [
          {
            "in": "path",
            "name": "shop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shop Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "notice_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Notice Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/qarz__interface__payment_notices_api__Accept"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Payment Notices Accept Api V1 Shops  Shop Id  Payment Notices  Notice Id  Accept Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Payment Notices.Accept"
      }
    },
    "/api/v1/shops/{shop_id}/payment-notices/{notice_id}/decline": {
      "post": {
        "operationId": "payment_notices_decline_api_v1_shops__shop_id__payment_notices__notice_id__decline_post",
        "parameters": [
          {
            "in": "path",
            "name": "shop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shop Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "notice_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Notice Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Decline"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Payment Notices Decline Api V1 Shops  Shop Id  Payment Notices  Notice Id  Decline Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Payment Notices.Decline"
      }
    },
    "/api/v1/shops/{shop_id}/payment-notices/{notice_id}/receipt": {
      "get": {
        "operationId": "payment_notices_receipt_api_v1_shops__shop_id__payment_notices__notice_id__receipt_get",
        "parameters": [
          {
            "in": "path",
            "name": "shop_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shop Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "notice_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Notice Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Payment Notices Receipt Api V1 Shops  Shop Id  Payment Notices  Notice Id  Receipt Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Payment Notices.Receipt"
      }
    },
...
  }
}